% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa04-ternarysystems.R
\name{addTernarySystem}
\alias{addTernarySystem}
\alias{addTernarySystem.ternarySystem}
\title{Add or update a ternary system to the list of available ternary systems (only for this session)}
\usage{
addTernarySystem(s, name, overwrite = FALSE)

\method{addTernarySystem}{ternarySystem}(s, name, overwrite = FALSE)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem-class}} object, 
such as obtained with 
\code{\link[ternaryplot]{getTernarySystem}}.}

\item{name}{Single character string. Name (label) of the ternary 
system to be added.}

\item{overwrite}{Single logical value. If \code{TRUE} (not the default), 
existing ternary system with the same name as in 
\code{name} will be overwritten.}
}
\value{
Nothing. Used for its side effect.
}
\description{
Add or update a ternary system to the list of available 
 ternary systems. Only valid for that R session and until 
 it is closed (i.e. not permanent effects). It only adds 
 the ternary system for the ternary-plot package.
}
\examples{

library( "ternaryplot" )

hypres <- getTernarySystem( s = "hypres" )

#   Reverse the axis direction
blrClock( s = hypres ) 
blrClock( s = hypres ) <- rep( FALSE, 3 ) 

addTernarySystem( s = hypres, name = "hypres_FFF" ) 

hypres_FFF <- getTernarySystem( s = "hypres_FFF" )

identical( hypres, hypres_FFF )
# [1] TRUE

#   Change the axis-variable attribution
blrNames( s = hypres_FFF )
blrNames( s = hypres_FFF ) <- c( "CLAY", "SILT", "SAND" ) 

addTernarySystem( s = hypres_FFF, name = "hypres_FFF", 
    overwrite = TRUE )  

hypres_FFF2 <- getTernarySystem( s = "hypres_FFF" )

identical( hypres_FFF, hypres_FFF2 )

#   Clean-up
deleteTernarySystem( name = "hypres_FFF" )
}
\seealso{
\code{\link[ternaryplot]{getTernarySystem}}, 
 \code{\link[ternaryplot]{listTernarySystem}} and 
 \code{\link[ternaryplot]{deleteTernarySystem}}.
 \code{\link[ternaryplot]{createTernarySystem}} should 
 be used to create/define a ternary system before adding 
 it to the list of available ternary systems.
}

