% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame.ternarySystem}
\alias{as.data.frame.ternarySystem}
\title{Extract the class vertices or the class centroids of a ternarySystem.}
\usage{
\method{as.data.frame}{ternarySystem}(x, what = "vertices", ...)
}
\arguments{
\item{x}{A \code{\link[ternaryplot]{ternarySystem-class}} object.
Contrary to many other functions in the 
ternaryplot-package, \code{x} should \emph{not} be a 
character string naming an existing \code{ternarySystem} 
(because the method 
\code{\link[base]{as.data.frame.character}} 
is already defined for other purposes).}

\item{what}{Single character string. What points should be extracted: 
the class vertices (\code{what = 'vertices'}, the default) 
or the class centroids (\code{what = 'centroids'})}

\item{\dots}{Additional parameters. Not used.}
}
\value{
A \code{\link[base]{data.frame}}. If \code{what} is 
 \code{'vertices'}, each row is a vertex, and the column 
 returned are \code{id}, as well as the 3 ternary variables 
 (as returned by \code{blrNames(s)}). If \code{what} is 
 \code{'centroids'}, the column 
 returned are \code{abbrev}, \code{name}, as well as the 
 3 ternary variables (as returned by \code{blrNames(s)}).
}
\description{
Extract the class vertices or the class centroids of a 
 \code{\link[ternaryplot]{ternarySystem-class}} object.
}
\examples{

library( "ternaryplot" )

hypres <- getTernarySystem( "hypres" )

vertices <- as.data.frame( hypres, what = 'vertices' ) 
vertices 

centroids <- as.data.frame( hypres, what = 'centroids' ) 
centroids

    # # Visualise the result:

    # ternaryStyle( "publication" )
    # ternaryPlot( hypres )
    # ternaryPoints( s = hypres, x = vertices, col = "darkred" ) 
    # ternaryPoints( s = hypres, x = centroids, col = "purple" )  
}

