\docType{methods}
\name{blrNames}
\alias{blrNames}
\alias{blrNames,ternarySystem-method}
\alias{blrNames<-}
\alias{blrNames<-,ternarySystem-method}
\title{Set or get the bottom-left-right axis names of a ternarySystem
 object}
\usage{
  \S4method{blrNames}{ternarySystem}( s, \dots )

  \S4method{blrNames}{ternarySystem}( s, \dots ) <- value
}
\arguments{
  \item{s}{A \code{\linkS4class{ternarySystem}} object.}

  \item{\dots}{Additional parameters passed to specific
  methods.}

  \item{value}{A vector of 3 character strings. Names of
  the bottom-left-right axis.}
}
\description{
  Set or get the bottom-left-right axis names of a
  \code{\linkS4class{ternarySystem}} object
}

