% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa02-ternaryplot-classes.R
\name{createTernarySystem}
\alias{createTernarySystem}
\title{Creates a ternarySystem-class object (the full definition of ternary plot and optionally classification).}
\usage{
createTernarySystem(ternaryGeometry = NULL, ternaryVariables = NULL,
  main = character(0), vertices = NULL, classes = NULL, scale = NULL,
  over = NULL, ...)
}
\arguments{
\item{ternaryGeometry}{See \code{\link[ternaryplot]{ternarySystem-class}}, 
\code{\link[ternaryplot]{ternaryGeometry-class}}, and 
\code{\link[ternaryplot]{createTernaryGeometry}}.}

\item{ternaryVariables}{See \code{\link[ternaryplot]{ternarySystem-class}}, 
\code{\link[ternaryplot]{ternaryVariables-class}}, and 
\code{\link[ternaryplot]{createTernaryVariables}}.}

\item{main}{See \code{\link[ternaryplot]{ternarySystem-class}}.}

\item{vertices}{See \code{\link[ternaryplot]{ternarySystem-class}}.}

\item{classes}{See \code{\link[ternaryplot]{ternarySystem-class}}.}

\item{scale}{See \code{\link[ternaryplot]{ternarySystem-class}}.}

\item{over}{See \code{\link[ternaryplot]{ternarySystem-class}}.}

\item{\dots}{Additional parameters passed to 
\code{\link[ternaryplot]{ternaryCheck}}}
}
\value{
A \code{\link[ternaryplot]{ternarySystem-class}}.
}
\description{
Creates a \code{\link[ternaryplot]{ternarySystem-class}} 
 object (the full definition of ternary plot and optionally 
 classification).
}
\examples{

library( "ternaryplot" ) 


## Default variables
tsy <- createTernarySystem() 
tsy 

## Custom variables
tsy2 <- createTernarySystem( 
    "ternaryVariables" = createTernaryVariables( 
        "blrNames"  = c( "CLAY", "SILT", "SAND" ), 
        "blrLabels" = c( "Clay [\%]", "Silt [\%]", "Sand [\%]" ) 
    ),  
    "ternaryGeometry" = createTernaryGeometry( ##
        "fracSum"   = 100, 
        "tlrAngles" = c( 45, 45, 90 ) # top, left, right angles
    ), 
    "main" = "My ternary plot" ) 

tsy2

}
\seealso{
\code{\link[ternaryplot]{ternarySystem-class}}.
}

