% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa06-ternary2xy.R
\name{deg2rad}
\alias{deg2rad}
\title{Converts angles expressed in degrees into radian.}
\usage{
deg2rad(angle)
}
\arguments{
\item{angle}{Single numeric. Angle in Degrees.}
}
\description{
Converts angles expressed in degrees into radian.
}

