% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa04-ternarysystems.R
\name{deleteTernarySystem}
\alias{deleteTernarySystem}
\title{Delete a ternary system from the list of available ternary systems (only for this session)}
\usage{
deleteTernarySystem(name)
}
\arguments{
\item{name}{Vector of character strings. Name of the ternary systems 
to be deleted from the list of available systems.}
}
\value{
Nothing. Used for its side effect.
}
\description{
Delete a ternary system from the list of available ternary 
 systems. Only valid for that R session (unless the 
 ternarySystem had been added during the session using 
 \code{\link[ternaryplot]{addTernarySystem}}, in which 
 case it disappears anyway at the end of the session). 
 Can only delete a ternary system for the ternary-plot 
 package.
}
\examples{

library( "ternaryplot" )

hypres <- getTernarySystem( s = "hypres" )

#   Reverse the axis direction
blrClock( s = hypres ) 
# [1] TRUE TRUE TRUE
blrClock( s = hypres ) <- rep( FALSE, 3 ) 

addTernarySystem( s = hypres, name = "hypres_FFF" ) 

"hypres_FFF" \%in\% listTernarySystem()[, "systemName" ]
# [1] TRUE

deleteTernarySystem( name = "hypres_FFF" ) 

"hypres_FFF" \%in\% listTernarySystem()[, "systemName" ]
# [1] FALSE
}
\seealso{
\code{\link[ternaryplot]{getTernarySystem}}, 
 \code{\link[ternaryplot]{listTernarySystem}} and 
 \code{\link[ternaryplot]{addTernarySystem}}.
}

