\docType{methods}
\name{fracSum}
\alias{fracSum}
\alias{fracSum,ternarySystem-method}
\alias{fracSum<-}
\alias{fracSum<-,ternarySystem-method}
\title{Set or get the sum of the three fractions of a ternarySystem
 object}
\usage{
  \S4method{fracSum}{ternarySystem}( x, \dots )

  \S4method{fracSum<-}{ternarySystem}( x, \dots ) <- value
}
\arguments{
  \item{x}{A \code{\linkS4class{ternarySystem}} object.}

  \item{\dots}{Additional parameters passed to specific
  methods.}

  \item{value}{Single numerical value. Sum of the three
  fractions of the ternarySystem.}
}
\description{
  Set or get the sum of the three fractions of a
  \code{\linkS4class{ternarySystem}} object
}

