% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa04-ternarysystems.R
\name{getAllTernarySystems}
\alias{getAllTernarySystems}
\title{INTERNAL: Fetch all the pre-defined ternary classification system}
\usage{
getAllTernarySystems()
}
\value{
A \code{\link[base]{list}} 
 \code{\link[ternaryplot]{ternarySystem-class}} objects.
}
\description{
INTERNAL: Fetch all the pre-defined ternary classification 
 system, as a list of 
 \code{\link[ternaryplot]{ternarySystem-class}}es.
}
\details{
The function is designed for internal use, in combination 
 with the argument \code{terSysEnvList} of the functions 
 \code{link[ternaryPlot]{tpPar}} or 
 \code{link[ternaryPlot]{getTpPar}}, as a flexible 
 mechanism to manage the list of 
 \code{\link[ternaryplot]{ternarySystem-class}} objects 
 used by the package.
}
\keyword{internal}

