\name{getTernarySystem}
\alias{getTernarySystem}
\title{Fetch a pre-defined ternary classification system}
\usage{
  getTernarySystem(s = "default")
}
\arguments{
  \item{s}{Single character string. Name of the ternary
  classification to be fetched.}
}
\value{
  A \code{ternarySystem} object.
}
\description{
  Fetch a pre-defined ternary classification system
}

