\name{listTernarySystem}
\alias{listTernarySystem}
\title{List all pre-defined ternary classification systems}
\usage{
  listTernarySystem()
}
\value{
  A vector of character strings, names of the pre-defined
  ternary classification systems
}
\description{
  List all pre-defined ternary classification systems
}

