% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa03-ternaryplot-classes-utility.R
\name{print.ternarySystem}
\alias{print.ternarySystem}
\title{Print the content of a ternarySystem object in a human readable format.}
\usage{
\method{print}{ternarySystem}(x, prefix = "", collapse = "; ", ...)
}
\arguments{
\item{x}{A \code{ternarySystem} object, as created with
 \code{\link[ternaryplot]{createTernarySystem}}.}

\item{prefix}{Single character string. Prefix used before the different
 items in \code{x} (intended for internal use, for example
 \code{prefix = "$ternaryGeometry"}).}

\item{collapse}{Single character string. Passed to
 \code{\link{paste}( ..., collapse )} when displaying the
 items' values.}

\item{\dots}{Additional parameters passed to specific methods (not
 used).}
}
\description{
Print the content of a \code{ternarySystem} object
 (S3-class) in a human readable format.
}

