% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa06-ternary2xy.R
\name{ternary2xy}
\alias{ternary2xy}
\alias{ternary2xy.character}
\alias{ternary2xy.ternarySystem}
\title{Converts ternary point-data into x-y coordinates}
\usage{
ternary2xy(s, ...)

\method{ternary2xy}{character}(s, ...)

\method{ternary2xy}{ternarySystem}(s, x, ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem}} object or a 
character string naming an pre-defined 
\code{\link[ternaryplot]{ternarySystem-class}}.}

\item{x}{A \code{\link[base]{data.frame}} or a \code{\link[base]{matrix}} 
with ternary point-data.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
Returns a \code{\link[base]{data.frame}} with point-data in 
 (columns) x and y.
}
\description{
Converts ternary point-data (bottom, left, right axis) into 
 x-y coordinates, according to the specification of a 
 \code{\link[ternaryplot]{ternarySystem-class}}
}

