\docType{methods}
\name{.ternaryAxisArrows}
\alias{.ternaryAxisArrows}
\alias{.ternaryAxisArrows,ternarySystem-method}
\title{INTERNAL: Draw axis' arrows and arrows' label marks on a
 triangle plot}
\usage{
  .ternaryAxisArrows(s, ...)
}
\arguments{
  \item{s}{A \code{\linkS4class{ternarySystem}} object.}

  \item{\dots}{Additional parameters passed to specific
  methods.}
}
\value{
  Invisibly returns a list of \code{data.frame} with the
  start and end points of the arrows segments
}
\description{
  INTERNAL: Draw axis' arrows and arrows' label marks on a
  triangle plot
}

