\docType{methods}
\name{.ternaryAxisArrowsBase}
\alias{.ternaryAxisArrowsBase}
\alias{.ternaryAxisArrowsBase,ternarySystem-method}
\title{INTERNAL. Calculates arrows segments start and end for ternary
 axis labels}
\usage{
  .ternaryAxisArrowsBase(s, ...)
}
\arguments{
  \item{s}{A \code{\linkS4class{ternarySystem}} object.}

  \item{\dots}{Additional parameters passed to specific
  methods.}
}
\description{
  INTERNAL. Calculates arrows segments start and end for
  ternary axis labels
}

