% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryBins.R
\name{ternaryBins}
\alias{ternaryBins}
\alias{ternaryBins.ternarySystem}
\title{Add bins (counts in ternary grid cells) to a ternary plot}
\usage{
ternaryBins(s, x, ...)

\method{ternaryBins}{ternarySystem}(s, x, grid = TRUE, noZero = TRUE, ...)
}
\arguments{
\item{s}{Either \itemize{
  \item A \code{\link[ternaryplot]{ternarySystem-class}}. 
    That \code{ternarySystem-class} must contain 
    a ternary classification. 
  \item A single character string, the name of an existing 
    pre-defined \code{\link[ternaryplot]{ternarySystem-class}}.
  \item A \code{\link[ternaryplot]{ternaryPolygons-class}}.
}}

\item{x}{A \code{\link[base]{data.frame}} or a 
\code{\link[base]{matrix}} containing point ternary data 
to be binned and counted. It should contain the 3 columns 
names given by \code{blrNames(s)}.}

\item{grid}{Single logical value. Set to \code{TRUE} (the default) 
to retrieve counts for a systematic ternary grid, and 
to \code{FALSE} to count the number of \code{x}-data 
points per class in \code{s} instead.}

\item{noZero}{Single logical value. If \code{TRUE} (the default), 
zero-counts values are not shown on the plot. This also 
makes the legend more readable.}

\item{\dots}{Additional parameters passed to 
\code{\link[ternaryplot]{ternaryPolygons}}. See in particular}
}
\value{
Invisibly returns the output of 
 \code{\link[ternaryplot]{ternaryCount}}.
}
\description{
Add bins (counts in ternary grid cells) to a ternary plot
}
\examples{

library( "ternaryplot" )

data( "textureDataset", package = "ternaryplot" ) 

# 'sandwitch' plotting of different plot components (with bins)
# ==========================================================

ternaryPlot( "hypres" )
s <- ternaryPlot( "hypres", axes = FALSE, grid = FALSE, 
    classes = FALSE ) 
cnts <- ternaryBins( s = s, x = textureDataset ) 
ternaryPolygons( s = s, bg = NA ) 
ternaryGrid( s = s )
ternaryBox( s = s )
ternaryAxis( s = s )
ternaryPoints( s = s, x = textureDataset ) 

head( attr( cnts, "data" ) ) 

# The same, all in one
# ==========================================================

ternaryPlot( s = "hypres", x = textureDataset, 
    type = c( "p", "c" ) )
}
\seealso{
\code{\link[ternaryplot]{ternaryPlot}} for plotting 
 ternary bins as a background of a ternary plot, and 
 \code{\link[ternaryplot]{ternaryCount}} for producing 
 the bins and count per bins but not plotting them.

 Package-options \code{ternaryPolygons.border.col} and \code{ternaryPolygons.bg.fun}  
 in \code{\link[ternaryplot]{tpPar}} are used internally.
}

