\docType{methods}
\name{ternaryBox}
\alias{ternaryBox}
\alias{ternaryBox,ternarySystem-method}
\title{Draw a frame (bounding-box) around an existing ternaryPlot}
\usage{
  ternaryBox(s, ...)
}
\arguments{
  \item{s}{A \code{\linkS4class{ternarySystem}} object.}

  \item{\dots}{Additional parameters passed to specific
  methods.}
}
\description{
  Draw a frame (bounding-box) around an existing
  ternaryPlot
}

