% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ternaryClasses.R
\name{ternaryClasses}
\alias{ternaryClasses}
\alias{ternaryClasses.character}
\alias{ternaryClasses.ternarySystem}
\title{Extract the ternary classification (class polygons) from a ternarySystem-object}
\usage{
ternaryClasses(s, ...)

\method{ternaryClasses}{character}(s, ...)

\method{ternaryClasses}{ternarySystem}(s, ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem}}-object (such
 as created with \code{\link[ternaryplot]{createTernarySystem}}
 or fetched with \code{\link[ternaryplot]{getTernarySystem}}),
 or a single character string naming a
 \code{\link[ternaryplot]{ternarySystem}}.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
A \code{ternaryPolygons}-object
 containing the coordinates of the class polygons
 extracted from \code{s}.
}
\description{
Extract the ternary classification (class polygons) from a
 ternarySystem-object.
}
\examples{

library( "ternaryplot" ) 

#   Fetch the definition of a ternary system
s <- ternaryPlot( s = "dummy" )

#   Create the ternary grid and convert it to 
#   sp SpatialPolygonsDataFrame
tc   <- ternaryClasses( s )

#   Plot the ternary system and the grid
ternaryPlot( s = s )
ternaryPolygons( s = tc, border = "darkred", lty = 3 ) 

# #   Alternatively:
tcSp <- ternary2SpatialPolygonsDataFrame( tc ) 
# ternaryPlot( s = s ) 
# sp::plot( tcSp, add = TRUE, border = "darkred", lty = 3 ) 
}
\seealso{
\code{\link[ternaryplot]{ternary2SpatialPolygonsDataFrame}}
 (to convert the grid into a
 \code{\link[sp]{SpatialPolygonsDataFrame}}), and
 \code{\link[ternaryplot]{ternaryPlot}} to plot the
 output.
}

