% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryClick.R
\name{ternaryClick}
\alias{ternaryClick}
\alias{ternaryClick.character}
\alias{ternaryClick.ternarySystem}
\title{Click on a ternary plot to get values of the ternary variables at that location.}
\usage{
ternaryClick(s, n = 1L, ...)

\method{ternaryClick}{character}(s, n = 1L, ...)

\method{ternaryClick}{ternarySystem}(s, n = 1L, ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem}} object or a 
character string naming an pre-defined 
\code{\link[ternaryplot]{ternarySystem-class}}.}

\item{n}{Single integer value. Number of clicks on the ternary plot.}

\item{\dots}{Additional parameters passed to 
\code{\link[graphics]{locator}}.}
}
\value{
Returns a \code{\link[base]{data.frame}} with ternary 
 point-data. With \code{n} rows and 3 columns, one per 
 ternary-variable (same names as \code{blrNames(s)}).
}
\description{
Click on a ternary plot to get values of the ternary 
 variables at that location.
}
\examples{

library( "ternaryplot" ) 

# Test ternaryClick with different ternary geometries
# ==========================================================

#   Number of clicks
n <- 2L

if( interactive() ){
    
    hypres <- getTernarySystem( "hypres" )
    
    blrClock( hypres )
    
    blrClock0 <- list(
        "TTT" = c(  TRUE,  TRUE,  TRUE ), 
        "FFF" = c( FALSE, FALSE, FALSE ), 
        "FTX" = c( FALSE,  TRUE,    NA ), 
        "TXF" = c(  TRUE,    NA, FALSE ) 
    )   
    
    ternaryStyle( "publication" )
    
    for( i in 1:length( blrClock0 ) ){
        message( paste( 
            ":::", 
            paste( blrClock0[[ i ]], collapse = ", " ), 
            ":::" 
        ) )  
        
        blrClock( hypres ) <- blrClock0[[ i ]] 
        
        ternaryPlot( hypres ) 
        
        print( ternaryClick( hypres, n = n, type = "p" ) ) 
        
        # flush.console()
    }   
}   

tpPar( reset = TRUE )
}
\seealso{
\code{\link[graphics]{locator}} and 
 \code{\link[ternaryplot]{xy2ternary}}. The first is used 
 to retrieve the x-y coordinates from the ternary plot 
 and the later to convert them into ternary-variables.
}

