\docType{methods}
\name{ternaryGrid}
\alias{ternaryGrid}
\alias{ternaryGrid,ternarySystem-method}
\title{Draw a triangular grid on a triangle plot}
\usage{
  ternaryGrid(s, ...)
}
\arguments{
  \item{s}{A \code{\linkS4class{ternarySystem}} object.}

  \item{\dots}{Additional parameters passed to specific
  methods.}
}
\value{
  Invisibly returns a list of \code{data.frame} with the
  start and end points of the grid segments
}
\description{
  Draw a triangular grid on a triangle plot
}

