\name{.ternaryGridBase}
\alias{.ternaryGridBase}
\alias{.ternaryGridBase.ternarySystem}
\title{INTERNAL. Calculates start- and end-points of grid segments in ternary coordinates}
\usage{
  .ternaryGridBase(s, ticks = FALSE, ...)

  \method{.ternaryGridBase}{ternarySystem} (s,
    ticks = FALSE, ...)
}
\arguments{
  \item{s}{A \code{\link[ternaryplot]{ternarySystem}}
  object.}

  \item{ticks}{Single logical. If \code{TRUE}, return
  values for axis' tick-marks. If \code{FALSE}, return
  values for a grid.}

  \item{\dots}{Additional parameters passed to specific
  methods.}
}
\description{
  INTERNAL. Calculates start- and end-points of grid
  segments in ternary coordinates
}
\keyword{internal}

