\docType{package}
\name{ternaryplot-package}
\alias{ternaryplot-package}
\title{Ternary plots and ternary classifications}
\description{
  Ternary plots and ternary classifications
}
\examples{
# Examples coming later
}
\author{
  Julien Moeys \email{Julien.Moeys@slu.se}

  Maintainer: Julien Moeys \email{Julien.Moeys@slu.se}
}
\keyword{package}

