% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryPolygons-class.R
\name{ternaryPolygons-class}
\alias{ternaryPolygons-class}
\title{Class to hold ternary polygons (ternary grids, ternary classification, ...)}
\description{
Class (\code{S3}) to hold ternary polygons (ternary grids, 
 ternary classification, ...). A 
 \code{ternaryPolygons}-class object is 
 a tagged \code{\link[base]{data.frame}} to which is 
 associated several \code{\link[base]{attributes}}.
}
\details{
Each row in the \code{data.frame} contains the ternary 
 coordinates of one vertex of one polygon. Polygons should 
 of course have at least 3 vertices and there can be 
 as many polygons as needed.


 The \code{S3} \code{\link[base]{class}} system is an 
 \emph{in}formal and lightweight class system.
}
\section{data.frame attributes}{


 Note: see \code{\link[base]{attr}} to retrieve or change 
 attributes.
 
 The \code{\link[base]{data.frame}} has the following 
 attributes:

\itemize{
 \item \bold{\code{ternarySystem}}: A 
   \code{\link[ternaryplot]{ternarySystem-class}} object. 
   The ternary system associated with the ternary polygons.
 \item \bold{\code{idCol}}: A single character string. 
   name of the column in the \code{\link[base]{data.frame}} 
   that contains the identifiers of the different polygons.
 \item \bold{\code{labels}}: A vector of character strings, 
   the labels to be displayed inside the polygons when 
   plotting them on a ternary plot. There should be 
   one label per polygon in the \code{\link[base]{data.frame}}, 
   and the labels should be given in the same order as 
   the polygons identifiers in the 
   \code{\link[base]{data.frame}}.
 \item \bold{\code{data}}: An optional 
   \code{\link[base]{data.frame}} containing additional 
   variables for each of the ternary polygon. It should 
   have at least one column, \code{idCol} the identifiers 
   of the different polygons (see above and below).
}
}

\section{data.frame columns}{

 
 The \code{\link[base]{data.frame}} has the following 
 column names:
 
\itemize{
 \item 3 columns named after the argument \code{blrNames} 
   in the \code{ternarySystem} attribute. Use 
   \code{\link[ternaryplot]{blrNames}( attr( x, "ternarySystem" ) )} 
   (where \code{x} is a \code{ternaryPolygons-class} object) 
   to fetch the current variable names.
 \item 1 column named after the attribute \code{idCol}. 
}
}
\seealso{
\itemize{
 \item \code{\link[ternaryplot]{createTernaryGrid}} 
   to create a regular ternary grid (of triangular 
   polygons) and \code{\link[ternaryplot]{ternaryClasses}} 
   to extract the ternary classification of a ternary system 
   as \code{ternaryPolygons-class}. 
 \item \code{\link[ternaryplot]{ternaryPolygons}} (a 
   function, not to be confused with 
   \code{ternaryPolygons-class}, the class definition) to 
   plot ternary polygons on a ternary triangle.
}
}

