\name{ternarySystemGet}
\alias{ternarySystemGet}
\title{Fetch a pre-defined ternary classification system}
\usage{
  ternarySystemGet(s = "default")
}
\arguments{
  \item{s}{Single character string. Name of the ternary
  classification to be fetched.}
}
\value{
  A \code{\linkS4class{ternarySystem}} object.
}
\description{
  Fetch a pre-defined ternary classification system
}

