\name{ternarySystemList}
\alias{ternarySystemList}
\title{List all pre-defined ternary classification systems}
\usage{
  ternarySystemList()
}
\value{
  A vector of character strings, names of the pre-defined
  ternary classification systems
}
\description{
  List all pre-defined ternary classification systems
}

