\docType{methods}
\name{.ternaryTicks}
\alias{.ternaryTicks}
\alias{.ternaryTicks,ternarySystem-method}
\title{Draw axis' ticks mark on a triangle plot}
\usage{
  .ternaryTicks(s, ...)
}
\arguments{
  \item{s}{A \code{\linkS4class{ternarySystem}} object.}

  \item{\dots}{Additional parameters passed to specific
  methods.}
}
\value{
  Invisibly returns a list of \code{data.frame} with the
  start and end points of the grid segments
}
\description{
  Draw axis' ticks mark on a triangle plot
}

