% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textureDataset.R
\name{textureDataset}
\alias{textureDataset}
\title{An example of soil texture dataset (generated, virtual, dummy example)}
\format{A data frame with 200 rows and 3 variables:
\itemize{
  \item \code{CLAY}: Clay content of the soil samples [\%].
  \item \code{SILT}: Silt content of the soil samples [\%].
  \item \code{SAND}: Sand content of the soil samples [\%].
  \item \code{GROUP}: Group to which each sample belong (dummy too).
}}
\usage{
textureDataset
}
\description{
An example of soil texture dataset (generated, virtual,
 dummy example).
}

