% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa01-ternaryplot-options.R
\docType{data}
\name{.tpParList}
\alias{.tpParList}
\title{Environment containng a backup copy of all ternaryplot package-options}
\format{An environment, where each item corresponds to an
 argument in \code{\link[ternaryplot]{tpPar}}.}
\usage{
.tpParList
}
\description{
Environment containing a backup copy of all ternaryplot
 package-options
}
\keyword{internal}

