% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa01-ternaryplot-options.R
\docType{data}
\name{tpParList}
\alias{tpParList}
\title{Environment containing all ternaryplot package-options}
\format{An environment, where each item corresponds to an
 argument in \code{\link[ternaryplot]{tpPar}}.}
\usage{
tpParList
}
\description{
Environment containing all ternaryplot package-options
}
\keyword{internal}

