\docType{methods}
\name{blrClock}
\alias{blrClock}
\alias{blrClock,ternarySystem-method}
\alias{blrClock<-}
\alias{blrClock<-,ternarySystem-method}
\title{Set or get the bottom-left-right orientation of a ternarySystem
 object}
\usage{
  \S4method{blrClock}{ternarySystem}( x, \dots )

  \S4method{blrClock<-}{ternarySystem}( x, \dots ) <- value
}
\arguments{
  \item{x}{A \code{\linkS4class{ternarySystem}} object.}

  \item{\dots}{Additional parameters passed to specific
  methods.}

  \item{value}{A vector of 3 logical values.
  Bottom-left-right orientation of a ternarySystem object.}
}
\description{
  Set or get the bottom-left-right orientation of a
  \code{\linkS4class{ternarySystem}} object
}

