% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa03-ternaryplot-classes-utility.R
\name{blrClock}
\alias{blrClock}
\alias{blrClock.ternaryGeometry}
\alias{blrClock.ternarySystem}
\alias{blrClock<-}
\alias{blrClock<-.ternaryGeometry}
\alias{blrClock<-.ternarySystem}
\title{Set or get the bottom-left-right orientation of a ternarySystem 
 object}
\usage{
blrClock(s, ...)

\method{blrClock}{ternarySystem}(s, ...)

\method{blrClock}{ternaryGeometry}(s, ...)

blrClock( s, ... ) <- value

\method{blrClock}{ternarySystem}(s, ...) <- value

\method{blrClock}{ternaryGeometry}(s, ...) <- value
}
\arguments{
\item{s}{Either \itemize{
  \item A \code{\link[ternaryplot]{ternarySystem-class}} 
    object, or 
  \item A \code{\link[ternaryplot]{ternaryGeometry-class}} 
    object.
}}

\item{value}{A vector of 3 logical values. Bottom-left-right orientation of 
a ternarySystem object.}

\item{\dots}{Additional parameters passed to 
\code{\link[ternaryplot]{ternaryCheck}}.}
}
\description{
Set or get the bottom-left-right orientation of a 
 \code{ternarySystem} object.
}
\examples{
#   See ?ternaryPlot for an example
}

