% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa03-ternaryplot-classes-utility.R
\name{blrLabels}
\alias{blrLabels}
\alias{blrLabels.ternarySystem}
\alias{blrLabels.ternaryVariables}
\alias{blrLabels<-}
\alias{blrLabels<-.ternarySystem}
\alias{blrLabels<-.ternaryVariables}
\title{Set or get the bottom-left-right axis labels of a ternarySystem 
 object}
\usage{
blrLabels(s, ...)

\method{blrLabels}{ternarySystem}(s, ...)

\method{blrLabels}{ternaryVariables}(s, ...)

blrLabels( s, ... ) <- value

\method{blrLabels}{ternarySystem}(s, ...) <- value

\method{blrLabels}{ternaryVariables}(s, ...) <- value
}
\arguments{
\item{s}{Either \itemize{
  \item A \code{\link[ternaryplot]{ternarySystem-class}} 
    object, or 
  \item A \code{\link[ternaryplot]{ternaryVariables-class}} 
    object.
}}

\item{value}{A vector of 3 character strings. Names of the 
bottom-left-right axis.}

\item{\dots}{Additional parameters passed to \code{\link[ternaryplot]{ternaryCheck}}.}
}
\value{
\code{blrLabels(s)} returns a vector of three character 
 strings, the names of the bottom, left and right 
 variable-labels.
}
\description{
Set or get the bottom-left-right axis labels of a 
 \code{\link[ternaryplot]{ternarySystem-class}} object.
}
\examples{
#   See ?ternaryPlot for an example
}

