% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa03-ternaryplot-classes-utility.R
\name{blrNames}
\alias{blrNames}
\alias{blrNames.ternarySystem}
\alias{blrNames.ternaryVariables}
\alias{blrNames<-}
\alias{blrNames<-.ternarySystem}
\alias{blrNames<-.ternaryVariables}
\title{Set or get the bottom-left-right axis names of a ternarySystem
 object}
\usage{
blrNames(s, ...)

\method{blrNames}{ternarySystem}(s, ...)

\method{blrNames}{ternaryVariables}(s, ...)

blrNames( s, ... ) <- value

\method{blrNames}{ternarySystem}(s, ...) <- value

\method{blrNames}{ternaryVariables}(s, ...) <- value
}
\arguments{
\item{s}{A \code{ternarySystem} object, as created with
 \code{\link[ternaryplot]{createTernarySystem}}, or a
 \code{ternaryVariables} object, as created with
 \code{\link[ternaryplot]{createTernaryVariables}}.}

\item{value}{A vector of 3 character strings. Names of the
 bottom-left-right axis.}

\item{\dots}{Additional parameters passed to \code{\link[ternaryplot]{ternaryCheck}}.}
}
\description{
Set or get the bottom-left-right axis names of a
 \code{ternarySystem} object.
}

