% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa02-ternaryplot-classes.R
\name{createTernaryGeometry}
\alias{createTernaryGeometry}
\title{Creates a ternaryGeometry-class object (settings of a ternary plot's geometry).}
\usage{
createTernaryGeometry(tlrAngles = c(60, 60, 60), blrClock = rep(TRUE, 3),
  fracSum = 100, ...)
}
\arguments{
\item{tlrAngles}{See \code{\link[ternaryplot]{ternaryGeometry-class}}.}

\item{blrClock}{See \code{\link[ternaryplot]{ternaryGeometry-class}}.}

\item{fracSum}{See \code{\link[ternaryplot]{ternaryGeometry-class}}.}

\item{\dots}{Additional parameters passed to
 \code{\link[ternaryplot]{ternaryCheck}}}
}
\value{
A a \code{\link[ternaryplot]{ternaryGeometry-class}}.
}
\description{
Creates a \code{\link[ternaryplot]{ternaryGeometry-class}}
 object (settings of a ternary plot's geometry).
}
\examples{

library( "ternaryplot" ) 


## Default geometry
tg <- createTernaryGeometry()
tg

## Custom geometry
tg <- createTernaryGeometry( 
    "fracSum"   = 1, 
    "blrClock"  = rep( FALSE, 3 ), 
    "tlrAngles" = c( 45, 45, 90 ) ) 
tg 

}
\seealso{
\code{\link[ternaryplot]{ternaryGeometry-class}}.
}

