% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createTernaryGrid.R
\name{createTernaryGrid}
\alias{createTernaryGrid}
\alias{createTernaryGrid.character}
\alias{createTernaryGrid.ternarySystem}
\title{Create a regular ternary grid. Base frame for binning ternary data.}
\usage{
createTernaryGrid(s, ...)

\method{createTernaryGrid}{character}(s, ...)

\method{createTernaryGrid}{ternarySystem}(s, n = 10, ...)
}
\arguments{
\item{s}{Either \itemize{
   \item A \code{\link[ternaryplot]{ternarySystem-class}} object.
   \item A character string naming an existing (pre-defined)
     \code{ternarySystem}.
 }}

\item{n}{Single integer value. Number of subdivisions of each
 axis (not the total number of cells!).}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
A \code{\link{ternaryPolygons-class}}.
}
\description{
Create a regular ternary grid. Base frame for binning ternary data.
}
\examples{

library( "ternaryplot" ) 

#   Fetch the definition of a ternary system
#   (which has not classes)
s <- ternaryPlot( s = "default" ) 


#   Create the ternary grid 
tg   <- createTernaryGrid( s )


#   Plot the ternary system and the grid
#   * Set the plot style (optional)
ternaryStyle( margin = TRUE ) 
#   * Base plot
ternaryPlot( s = s )
#   * Grid overlay
ternaryPolygons( s = tg, border = "darkred", lty = 3 ) 


# #   Alternatively:
tgSp <- ternary2SpatialPolygons( tg ) 

# #   sp overlay
# ternaryPlot( s = s )
# sp::plot( tgSp, add = TRUE, border = "darkred", lty = 3 ) 
}
\seealso{
\code{\link[ternaryplot]{ternaryPolygons-class}}.
}

