% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createTernaryGrid.R
\name{createTernaryGrid}
\alias{createTernaryGrid}
\alias{createTernaryGrid.character}
\alias{createTernaryGrid.ternarySystem}
\title{Create a regular ternary grid. Base frame for binning ternary data.}
\usage{
createTernaryGrid(s, ...)

\method{createTernaryGrid}{character}(s, ...)

\method{createTernaryGrid}{ternarySystem}(s, n = 10, ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem}}-object (such
 as created with \code{\link[ternaryplot]{createTernarySystem}}
 or fetched with \code{\link[ternaryplot]{getTernarySystem}}),
 or a single character string naming a
 \code{\link[ternaryplot]{ternarySystem}}.}

\item{n}{Single integer value. Number of grid cells (not grid
 nodes)}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
A ternary grid (of polygons), with class \code{ternaryPolygons}.
}
\description{
Create a regular ternary grid. Base frame for binning ternary data.
}
\examples{

library( "ternaryplot" ) 

#   Fetch the definition of a ternary system
#   (which has not classes)
s <- ternaryPlot( s = "default" ) 


#   Create the ternary grid and convert it to 
#   sp SpatialPolygonsDataFrame
tg   <- createTernaryGrid( s )


#   Plot the ternary system and the grid
ternaryPlot( s = s )
ternaryPolygons( s = tg, border = "darkred", lty = 3 ) 


# #   Alternatively:
tgSp <- ternary2SpatialPolygons( tg ) 
# ternaryPlot( s = s )
# sp::plot( tgSp, add = TRUE, border = "darkred", lty = 3 ) 
}
\seealso{
\code{\link[ternaryplot]{ternary2SpatialPolygons}}
 (to convert the grid into a
 \code{\link[sp]{SpatialPolygonsDataFrame}}), and
 \code{\link[ternaryplot]{ternaryPlot}} to plot the
 output.
}

