% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa02-ternaryplot-classes.R
\name{createTernaryVariables}
\alias{createTernaryVariables}
\title{Creates a ternaryVariables-class object (names and labels of ternary data and ternary plot)}
\usage{
createTernaryVariables(blrNames = c("_bottom_", "_left_", "_right_"),
  blrLabels = rep(NA_character_, 3), ...)
}
\arguments{
\item{blrNames}{See \code{\link[ternaryplot]{ternaryVariables-class}}.}

\item{blrLabels}{See \code{\link[ternaryplot]{ternaryVariables-class}}.}

\item{\dots}{Additional parameters passed to 
\code{\link[ternaryplot]{ternaryCheck}}}
}
\value{
A \code{\link[ternaryplot]{ternaryVariables-class}}.
}
\description{
Creates a ternaryVariables-class object (names and labels 
 of ternary data and ternary plot)
}
\examples{

library( "ternaryplot" ) 


## Default variables
tv <- createTernaryVariables() 
tv 

## Custom variables
tv2 <- createTernaryVariables( 
    blrNames  = c( "CLAY", "SILT", "SAND" ), 
    blrLabels = c( "Clay [\%]", "SILT [\%]", "SAND [\%]" ) ) 
tv2 

}
\seealso{
\code{\link[ternaryplot]{ternaryVariables-class}}.
}

