% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa03-ternaryplot-classes-utility.R
\name{fracSum}
\alias{fracSum}
\alias{fracSum.ternaryGeometry}
\alias{fracSum.ternarySystem}
\alias{fracSum<-}
\alias{fracSum<-.ternaryGeometry}
\alias{fracSum<-.ternarySystem}
\title{Set or get the sum of the three fractions of a ternarySystem
 object}
\usage{
fracSum(s, ...)

\method{fracSum}{ternarySystem}(s, ...)

\method{fracSum}{ternaryGeometry}(s, ...)

fracSum( s, ... ) <- value

\method{fracSum}{ternarySystem}(s, ...) <- value

\method{fracSum}{ternaryGeometry}(s, ...) <- value
}
\arguments{
\item{s}{A \code{ternarySystem} object, as created with
 \code{\link[ternaryplot]{createTernarySystem}}, or a
 \code{ternaryGeometry} object, as created with
 \code{\link[ternaryplot]{createTernaryGeometry}}.}

\item{value}{Single numerical value. Sum of the three fractions of the
 ternarySystem.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\description{
Set or get the sum of the three fractions of a
 \code{ternarySystem} object.
}

