% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa03-ternaryplot-classes-utility.R
\name{fracSum}
\alias{fracSum}
\alias{fracSum.ternaryGeometry}
\alias{fracSum.ternarySystem}
\alias{fracSum<-}
\alias{fracSum<-.ternaryGeometry}
\alias{fracSum<-.ternarySystem}
\title{Set or get the sum of the three fractions of a ternarySystem 
 object}
\usage{
fracSum(s, ...)

\method{fracSum}{ternarySystem}(s, ...)

\method{fracSum}{ternaryGeometry}(s, ...)

fracSum(s, ... ) <- value

\method{fracSum}{ternarySystem}(s, ...) <- value

\method{fracSum}{ternaryGeometry}(s, ...) <- value
}
\arguments{
\item{s}{Either \itemize{
  \item A \code{\link[ternaryplot]{ternarySystem-class}} 
    object, or 
  \item A \code{\link[ternaryplot]{ternaryGeometry-class}} 
    object.
}}

\item{value}{Single numerical value. Sum of the three fractions of the 
ternarySystem.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
\code{fracSum(s)} returns a single numeric value, the sum 
 of the three variables (1 if they are fractions and 100 
 if they are percentages).
}
\description{
Set or get the sum of the three fractions of a 
 \code{\link[ternaryplot]{ternarySystem-class}} object.
}

