% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa04-ternarysystems.R
\name{getTernarySystem}
\alias{getTernarySystem}
\title{Fetch a pre-defined ternary classification system}
\usage{
getTernarySystem(s = "default")
}
\arguments{
\item{s}{Single character string. Name of the ternary classification to
 be fetched.}
}
\value{
A \code{ternarySystem} object.
}
\description{
Fetch a pre-defined ternary classification system
}

