% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa04-ternarysystems.R
\name{getTernarySystem}
\alias{getTernarySystem}
\title{Fetch a pre-defined ternary classification system}
\usage{
getTernarySystem(s = "default")
}
\arguments{
\item{s}{Single character string. Name of the ternary classification to 
be fetched.}
}
\value{
A \code{\link[ternaryplot]{ternarySystem-class}} object.
}
\description{
Fetch a pre-defined ternary classification system
}
\examples{

library( "ternaryplot" )

hypres <- getTernarySystem( s = "hypres" )

class( hypres )

#   ternaryPlot( hypres )
}
\seealso{
\code{\link[ternaryplot]{listTernarySystem}}, 
 \code{\link[ternaryplot]{addTernarySystem}} and 
 \code{\link[ternaryplot]{deleteTernarySystem}}.
}

