% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary2SpatialPoints.R
\name{ternary2SpatialPoints}
\alias{ternary2SpatialPoints}
\alias{ternary2SpatialPoints.character}
\alias{ternary2SpatialPoints.ternarySystem}
\title{Converts ternary*-class objects to SpatialPoints}
\usage{
ternary2SpatialPoints(s, ...)

\method{ternary2SpatialPoints}{ternarySystem}(s, x = NULL,
  what = "vertices", ...)

\method{ternary2SpatialPoints}{character}(s, what = "vertices", ...)
}
\arguments{
\item{s}{Either \itemize{
  \item A \code{\link[ternaryplot]{ternarySystem-class}} 
    object, or 
  \item A single character string naming an existing  
    \code{\link[ternaryplot]{ternaryVariables-class}}.
}  
If \code{x} is \code{NULL}, them the vertices or the centroids 
of the ternary classification in \code{s} will be retrieved 
(see argument \code{what}).}

\item{x}{A \code{\link{data.frame}} containing the ternary data to 
be converted to a \code{\link[sp]{SpatialPoints}}-object.}

\item{what}{Single character string. When \code{x} is missing, what 
points should be extracted: the class vertices 
(\code{what = 'vertices'}, the default) or the class 
centroids (\code{what = 'centroids'})}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
A \code{\link[sp]{SpatialPoints}}
}
\description{
Converts ternary*-class objects to 
 \code{\link[sp]{SpatialPoints}}
}
\examples{

library( "ternaryplot" ) 

#   Plot a ternary diagram with some classification
#   * Set the style
ternaryStyle( "margin" = TRUE ) 
#   * Plot diagram
s <- ternaryPlot( "hypres" ) 

#   Extract classification vertices and centroids 
pts1 <- ternary2SpatialPoints( s )
pts2 <- ternary2SpatialPoints( s, what = "centroids" )

#   Add to the existing ternary plot
sp::plot( pts1, add = TRUE, col = "darkred", pch = 1 ) 
sp::plot( pts2, add = TRUE, col = "darkblue", pch = 3 ) 

tpPar( reset = TRUE )
}

