% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ternary2SpatialPolygons.R
\name{ternary2SpatialPolygons}
\alias{ternary2SpatialPolygons}
\alias{ternary2SpatialPolygons.ternaryPolygons}
\title{Converts ternary*-class objects to SpatialPolygonsDataFrame}
\usage{
ternary2SpatialPolygons(x, ...)

\method{ternary2SpatialPolygons}{ternaryPolygons}( x, ... )
}
\arguments{
\item{x}{A ternary*-class object.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
A \code{\link[sp]{SpatialPolygonsDataFrame}}
}
\description{
Converts ternary*-class objects to \code{\link[sp]{SpatialPolygonsDataFrame}}
}
\examples{
#   See ?createTernaryGrid or ?ternaryClasses for examples
}

