% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary2SpatialPolygons.R
\name{ternary2SpatialPolygons}
\alias{ternary2SpatialPolygons}
\alias{ternary2SpatialPolygons.character}
\alias{ternary2SpatialPolygons.ternaryPolygons}
\alias{ternary2SpatialPolygons.ternarySystem}
\title{Converts ternary*-class objects to SpatialPolygons}
\usage{
ternary2SpatialPolygons(s, ...)

\method{ternary2SpatialPolygons}{ternaryPolygons}(s, ...)

\method{ternary2SpatialPolygons}{ternarySystem}(s, ...)

\method{ternary2SpatialPolygons}{character}(s, ...)
}
\arguments{
\item{s}{Either \itemize{
  \item A \code{\link[ternaryplot]{ternaryPolygons-class}} 
    object, or 
  \item A \code{\link[ternaryplot]{ternarySystem-class}} 
    object, or 
  \item A single character string naming an existing  
    \code{\link[ternaryplot]{ternaryVariables-class}}.
}}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
A \code{\link[sp]{SpatialPolygons}} object.
}
\description{
Converts ternary*-class objects to \code{\link[sp]{SpatialPolygons}}
}
\examples{
#   See ?createTernaryGrid or ?ternaryClasses for examples
}

