% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ternarygrid.R
\name{ternary2SpatialPolygonsDataFrame}
\alias{ternary2SpatialPolygonsDataFrame}
\alias{ternary2SpatialPolygonsDataFrame.ternaryPolygons}
\title{Converts ternary*-class objects to SpatialPolygonsDataFrame}
\usage{
ternary2SpatialPolygonsDataFrame(x, ...)

\method{ternary2SpatialPolygonsDataFrame}{ternaryPolygons}( x, ... )
}
\arguments{
\item{x}{A ternary*-class object.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
A \code{\link[sp]{SpatialPolygonsDataFrame}}
}
\description{
Converts ternary*-class objects to \code{\link[sp]{SpatialPolygonsDataFrame}}
}

