% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa06-ternary2xy.R
\name{ternary2xy}
\alias{ternary2xy}
\alias{ternary2xy.character}
\alias{ternary2xy.ternaryData}
\alias{ternary2xy.ternarySystem}
\title{Converts ternary point-data into x-y coordinates}
\usage{
ternary2xy(s, ...)

\method{ternary2xy}{character}(s, ...)

\method{ternary2xy}{ternaryData}(s, ...)

\method{ternary2xy}{ternarySystem}(s, x, ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem}} object or a
 character string naming an pre-defined \code{ternarySystem},
 or a \code{\link[ternaryplot]{ternaryData}} object. For the
 latter, if \code{x} is missing, \code{s} will also be used
 for \code{x}}

\item{x}{A \code{\link[base]{data.frame}} or a \code{\link[base]{matrix}}
 with ternary point-data. Can be missing if \code{s} is a
 \code{\link[ternaryplot]{ternaryData}} object.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
Returns a \code{\link[base]{data.frame}} with point-data in
 (columns) x and y.
}
\description{
Converts ternary point-data (bottom, left, right axis) into
 x-y coordinates, according to the specification of a
 \code{ternarySystem}
}

