% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{ternaryArrows}
\alias{ternaryArrows}
\alias{ternaryArrows.ternarySystem}
\title{Add Arrows to a ternary plot}
\usage{
ternaryArrows(s, from, to, ...)

\method{ternaryArrows}{ternarySystem}(s, from, to, .plot = TRUE, ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem}} object, or a
 character string naming a pre-defined \code{ternarySystem}.
 If missing, set to \code{default}.}

\item{from}{A \code{\link[base]{data.frame}} or a
 \code{\link[base]{matrix}} containing the ternary
 coordinates of points *from* which to draw the arrows.
 Each row is an arrow, and the columns must correspond
 to \code{blrNames(s)} (variable names for the bottom,
 left and right axis).}

\item{to}{A \code{\link[base]{data.frame}} or a
 \code{\link[base]{matrix}} containing the ternary
 coordinates of points *to* which to draw the arrows.
 Each row is an arrow, and the columns must correspond
 to \code{blrNames(s)} (variable names for the bottom,
 left and right axis).}

\item{.plot}{Single logical value. Set to \code{FALSE} if you don't want
 to plot the graphical element and simply returns them as
 x-y coordinates (or \code{Spatial*} objects if \code{sp} is
 set to \code{TRUE} in \code{\link{tpPar}}).}

\item{\dots}{Additional parameters passed to \code{\link[graphics]{arrows}}.}
}
\value{
Invisibly returns the graphical element as x-y coordinates or
 a \code{Spatial*} objects (see \code{.plot}).
}
\description{
Add Arrows to a ternary plot
}
\seealso{
\code{\link[graphics]{arrows}}.
}

