% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{.ternaryAxisArrows}
\alias{.ternaryAxisArrows}
\alias{.ternaryAxisArrows.ternarySystem}
\title{INTERNAL: Draw axis' arrows and arrows' label marks on a
 triangle plot}
\usage{
.ternaryAxisArrows(s, ...)

\method{.ternaryAxisArrows}{ternarySystem}(s, ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem}} object.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
Invisibly returns a list of \code{data.frame}
 with the start and end points of the arrows segments
}
\description{
INTERNAL: Draw axis' arrows and arrows' label marks on a
 triangle plot
}
\keyword{internal}

