% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{.ternaryAxisArrowsBase}
\alias{.ternaryAxisArrowsBase}
\alias{.ternaryAxisArrowsBase.ternarySystem}
\title{INTERNAL. Calculates arrows segments start and end for ternary
 axis labels}
\usage{
.ternaryAxisArrowsBase(s, ...)

\method{.ternaryAxisArrowsBase}{ternarySystem}(s, ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem}} object.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\description{
INTERNAL. Calculates arrows segments start and end for ternary
 axis labels
}
\keyword{internal}

