% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{ternaryBox}
\alias{ternaryBox}
\alias{ternaryBox.ternarySystem}
\title{Draw a Box around a ternary plot}
\usage{
ternaryBox(s, ...)

\method{ternaryBox}{ternarySystem}(s, .plot = TRUE, ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem}} object.}

\item{.plot}{Single logical value. Set to \code{FALSE} if you don't want
 to plot the graphical element and simply returns them as
 x-y coordinates (or \code{Spatial*} objects if \code{sp} is
 set to \code{TRUE} in \code{\link{tpPar}}).}

\item{\dots}{Additional parameters passed to \code{\link[graphics]{polygon}}.
 You can for instance set \code{border} for the color of the
 box outline, \code{col}, for the fill-color of the box,
 \code{lwd} (outline thickness) or \code{lty} (line-type).}
}
\value{
Invisibly returns the graphical element as x-y coordinates or
 a \code{Spatial*} objects (see \code{.plot}).
}
\description{
Draw a Box around a ternary plot
}
\seealso{
\code{\link[graphics]{box}}.
}

