% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryBox.R
\name{ternaryBox}
\alias{ternaryBox}
\alias{ternaryBox.ternarySystem}
\title{Draw a Box around a ternary plot}
\usage{
ternaryBox(s, ...)

\method{ternaryBox}{ternarySystem}(s, bg = NA, ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem}} object.}

\item{bg}{Single character string, representing a colour. Colour 
of the box filling (background). Default is \code{NA}, 
no fill-colour.}

\item{\dots}{Additional parameters passed to \code{\link[graphics]{polygon}}.
You can for instance set \code{border} for the color of the 
box outline, \code{col}, for the fill-color of the box, 
\code{lwd} (outline thickness) or \code{lty} (line-type).}
}
\value{
Invisibly returns the graphical element as x-y coordinates.
}
\description{
Draw a Box around a ternary plot
}
\seealso{
\code{\link[graphics]{box}}.
}

