% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa02-ternaryplot-classes.R
\name{ternaryCheck}
\alias{ternaryCheck}
\alias{ternaryCheck.ternaryGeometry}
\alias{ternaryCheck.ternarySystem}
\alias{ternaryCheck.ternaryVariables}
\title{Check the validity of ternary*-class objects}
\usage{
ternaryCheck(x, ...)

\method{ternaryCheck}{ternaryGeometry}(x, ...)

\method{ternaryCheck}{ternaryVariables}(x, ...)

\method{ternaryCheck}{ternarySystem}(x, ...)
}
\arguments{
\item{x}{A ternary*-class object.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\description{
Check the validity of ternary*-class objects
}
\seealso{
Arguments \code{onFailure} and \code{okClock} in
 \code{\link[ternaryplot]{getTpPar}} (package options).
}

