% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa02-ternaryplot-classes.R, R/aa05-ternarydata.R
\name{ternaryCheck}
\alias{ternaryCheck}
\alias{ternaryCheck.data.frame}
\alias{ternaryCheck.ternaryGeometry}
\alias{ternaryCheck.ternarySystem}
\alias{ternaryCheck.ternaryVariables}
\title{Check the validity of ternary*-class objects}
\usage{
ternaryCheck(x, ...)

\method{ternaryCheck}{ternaryGeometry}(x, ...)

\method{ternaryCheck}{ternaryVariables}(x, ...)

\method{ternaryCheck}{ternarySystem}(x, ...)

\method{ternaryCheck}{data.frame}(x, s, testRange, testSum, ...)
}
\arguments{
\item{x}{A ternary*-class object, or a \code{\link{data.frame}}
 containing ternary data.}

\item{s}{If \code{x} is a \code{\link{data.frame}}, a
 \code{ternarySystem}-object, such as obtained with
 \code{\link[ternaryplot]{getTernarySystem}} or output
 by \code{\link[ternaryplot]{ternaryPlot}}, that defines
 how the ternary data in \code{x} should be (variable
 names and sum of 3 ternary fractions). Not needed if
 \code{x} is not a \code{\link{data.frame}}.}

\item{testRange}{Single logical. Test if the range of fraction is between 0 and
 the expected sum of fractions (1 or 100).}

\item{testSum}{Single logical. Test if the sum of the 3 fractions is equal to
 the expected sum of fractions (1 or 100).}

\item{\dots}{Additional parameters passed to specific methods.}
}
\description{
Check the validity of ternary*-class objects
}
\seealso{
Arguments \code{onFailure} and \code{okClock} in
 \code{\link[ternaryplot]{getTpPar}} (package options).
}

