% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa02-ternaryplot-classes.R
\name{ternaryCheck}
\alias{ternaryCheck}
\alias{ternaryCheck.ternaryGeometry}
\alias{ternaryCheck.ternarySystem}
\alias{ternaryCheck.ternaryVariables}
\title{Check the validity of ternary*-class objects and ternary data.}
\usage{
ternaryCheck(s, ...)

\method{ternaryCheck}{ternaryGeometry}(s, ...)

\method{ternaryCheck}{ternaryVariables}(s, ...)

\method{ternaryCheck}{ternarySystem}(s, x = NULL, testRange, testSum, ...)
}
\arguments{
\item{s}{Either \itemize{
  \item a \code{\link[ternaryplot]{ternarySystem-class}} 
    (needed when \code{x} is not \code{NULL}), or 
  \item a \code{\link[ternaryplot]{ternaryVariables-class}}, or 
  \item a \code{\link[ternaryplot]{ternaryGeometry-class}} 
    object. 
}  
Object whose validity is to be evaluated. When \code{x} 
is not \code{NULL}, it is the validity of the data in 
\code{x} that is checked, not that of \code{s}.}

\item{x}{If not \code{NULL}, a \code{\link[base]{data.frame}} or 
a \code{\link[base]{matrix}} containing ternary data to 
be checked.}

\item{testRange}{Single logical. Test if the range of fraction is between 
0 and the expected sum of fractions (1 or 100\%). Only 
when}

\item{testSum}{Single logical. Test if the sum of the 3 fractions is 
equal to the expected sum of fractions (1 or 100\%).}

\item{\dots}{Not used.}
}
\value{
Whether an output is returned depends on the package 
 parameter \code{onFailure} (see 
 \code{\link[ternaryplot]{tpPar}}; the function to be 
 used when a problem is found). If \code{onFailure} is 
 \code{\link[base]{stop}}, then an error is raised and 
 the evaluation stops. If it is a 
 \code{\link[base]{warning}} or a 
 \code{\link[base]{message}} the evaluation continues 
 and \code{TRUE} is returned when no problem was found, 
 and \code{FALSE} when one or several problems were found.
}
\description{
Check the validity of ternary*-class objects and ternary data.
}
\seealso{
Arguments \code{onFailure} and \code{okClock} in 
 \code{\link[ternaryplot]{getTpPar}} (package options).
}

