% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa02-ternaryplot-classes.R
\name{ternaryGeometry-class}
\alias{ternaryGeometry-class}
\title{Class to hold geometrical settings of a ternary plot (graphics)}
\description{
Class (\code{S3}) to hold geometrical settings of a ternary 
 plot (graphics). A \code{ternaryGeometry}-class object is 
 a tagged \code{\link[base]{list}} of several 
 objects/items (see below) that describes the geometry 
 of a ternary plot (ternary diagram, ternary triangle). 
 It describes the triangle's angles and the direction of 
 its axis (see below). A 'tagged' list means a list of 
 the form \code{list( "tag1" = value1, "tag2" = value2, 
 ... )}.
}
\details{
The \code{S3} \code{\link[base]{class}} system is an 
 \emph{in}formal and lightweight class system.
}
\section{List objects/items}{

\itemize{
 \item \bold{\code{tlrAngles}}: Vector of 3 numerical 
   values. The angles of the \bold{t}op, \bold{l}eft, and 
   \bold{r}ight vertices of the triangle. In degrees. 
   The sum of the 3 angles should thus always be 180.
 \item \bold{\code{blrClock}}: Vector of 3 logical/boolean 
   values (one \code{\link[base]{NA}}-value is allowed). 
   The direction of the \bold{b}ottom, \bold{l}eft, and 
   \bold{r}ight axis. \code{TRUE} when an axis is clockwise 
   (i.e. the values on the axis increase clockwise), 
   \code{FALSE} when it is counter-clockwise (i.e. the 
   values on the axis increase counter-clockwise) and 
   \code{\link[base]{NA}} when the values increase when 
   moving away from the axis (inside the triangle; in with 
   lines parallel to the axis being isolines, for that 
   axis). Possible combinations are \code{c(TRUE,TRUE,TRUE)}, 
   \code{c(FALSE,FALSE,FALSE)}, \code{c(FALSE,TRUE,NA)} 
   and \code{c(TRUE,NA,FALSE)}.
 \item \bold{\code{fracSum}}: Single numerical value. 
   Sum of the 3 fractions. Possible values are \code{1} 
   (when values/units are fractions) or \code{100} (when 
   values/units are percentages).
}
}
\seealso{
\itemize{
 \item \code{\link[ternaryplot]{ternarySystem-class}} 
  (full ternary data and plot definition), that builds 
   on \code{ternaryGeometry-class}, 
   \code{\link[ternaryplot]{ternaryVariables-class}} and 
   other settings.
 \item \code{\link[ternaryplot]{createTernaryGeometry}} 
   for setting up a \code{ternaryGeometry-class} object.
 \item \code{\link[ternaryplot]{tlrAngles}} for fetching 
   or setting the geometry's angle, 
   \code{\link[ternaryplot]{blrClock}} for fetching 
   or setting the geometry axis' directions and 
   \code{\link[ternaryplot]{fracSum}} for fetching 
   or setting the geometry's sum of fractions.
 \item \code{\link[ternaryplot]{ternaryGeometry}} 
   for fetching or changing the geometry of a 
   \code{ternarySystem-class} object.
 \item \code{\link[ternaryplot]{ternaryCheck}}  
   for checking the validity of a \code{ternaryGeometry-class} 
   object.
}
}

