% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa03-ternaryplot-classes-utility.R
\name{ternaryGeometry}
\alias{ternaryGeometry}
\alias{ternaryGeometry.ternarySystem}
\alias{ternaryGeometry<-.ternarySystem}
\title{Set or get the ternaryGeometry of a ternarySystem object}
\usage{
ternaryGeometry(s, ...)

\method{ternaryGeometry}{ternarySystem}(s, ...)

\method{ternaryGeometry}{ternarySystem}( s, ... ) <- value
}
\arguments{
\item{s}{A \code{ternarySystem} object, as created with
 \code{\link[ternaryplot]{createTernarySystem}}.}

\item{value}{A \code{ternaryGeometry} object, as created with
 \code{\link[ternaryplot]{createTernaryGeometry}}.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\description{
Set or get the ternaryGeometry of a \code{ternarySystem} object.
}

