\docType{methods}
\name{.ternaryGridBase}
\alias{.ternaryGridBase}
\alias{.ternaryGridBase,ternarySystem-method}
\title{Internal. Calculates grid segments start and end ternary coordinates}
\usage{
  .ternaryGridBase(s, ticks = FALSE, ...)
}
\arguments{
  \item{s}{A \code{\linkS4class{ternarySystem}} object.}

  \item{ticks}{Single logical. If \code{TRUE}, return
  values for axis' tick-marks. If \code{FALSE}, return
  values for a grid.}

  \item{\dots}{Additional parameters passed to specific
  methods.}
}
\description{
  Internal. Calculates grid segments start and end ternary
  coordinates
}

