% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{.ternaryGridBase}
\alias{.ternaryGridBase}
\alias{.ternaryGridBase.ternarySystem}
\title{INTERNAL. Calculates start- and end-points of grid segments (tick-marks, ...) in ternary coordinates}
\usage{
.ternaryGridBase(s, type = "grid", ...)

\method{.ternaryGridBase}{ternarySystem}(s, type = "grid", ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem}} object.}

\item{type}{Single character string. If \code{"grid"}, return values
 the grid in a ternary plot, if \code{"ticks"} return values
 for axis' tick-marks and if \code{"tickLabels"} returns
 values for axis' tick-marks' labels.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\description{
INTERNAL. Calculates start- and end-points of grid segments (tick-marks, ...) in ternary coordinates
}
\keyword{internal}

