% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{.ternaryLims}
\alias{.ternaryLims}
\alias{.ternaryLims.character}
\alias{.ternaryLims.ternarySystem}
\title{INTERNAL: Find optimal axis limits for a ternary plot.}
\usage{
.ternaryLims(s, ...)

\method{.ternaryLims}{character}(s, ...)

\method{.ternaryLims}{ternarySystem}(s, x, ...)
}
\arguments{
\item{s}{Either a character string naming the ternary classification
 system to be used (if pre-defined) or a
 \code{\link[ternaryplot]{ternarySystem}}, instead of \code{x}.}

\item{x}{A \code{\link[base]{data.frame}} or a \code{\link[base]{matrix}}
 containing point ternary data (x-y-x) to be ploted on the graph.
 It should contain the 3 columns names given in \code{s};}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
Returns a \code{\link[base]{data.frame}} with 3 columns (bottom,
 left and right variables) and 2 rows: the minimun value and the
 maximum value of each variable.
}
\description{
INTERNAL: Find optimal axis limits for a ternary plot.
}
\keyword{internal}

