% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa00-ternaryplot-package.R
\docType{package}
\name{ternaryplot-package}
\alias{ternaryplot-package}
\title{Infrastructure for ternary plots, ternary classifications and ternary binning (S3-classes).}
\description{
Infrastructure for ternary plots, ternary classifications and ternary binning (S3-classes).
}
\details{
Infrastructure for ternary plots, ternary 
   classifications and ternary binning (with S3-classes). 
   Designed primarily to support the development of soil 
   texture plots (soil texture diagrams), but should be 
   a useful infrastructure for other ternary plots as well, 
   and especially for ternary classifications. Plots fully 
   compatible with the graphics-package. 'ternaryplot' 
   attempts to respect as much as possible (a) the plot 
   layout of the graphics-package, (b) the size and location 
   of plot-items such as axis tick marks and (c) the name 
   of parameters in the graphics-package. Function names and 
   functions arguments are as much as possible similar to 
   those in the graphics-package. Several ternary plot 
   geometries are supported, but all are based on a single 
   triangle with a horizontal base.
}
\examples{
 # Examples coming later

}
\author{
Julien Moeys \email{Julien.Moeys@slu.se} 

Maintainer: Julien Moeys \email{Julien.Moeys@slu.se}
}
\keyword{package}

