% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryPoints.R
\name{ternaryPoints}
\alias{ternaryPoints}
\alias{ternaryPoints.ternarySystem}
\title{Add points to a ternary plot}
\usage{
ternaryPoints(s, ...)

\method{ternaryPoints}{ternarySystem}(s, x, ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem}} object.}

\item{x}{A \code{\link[base]{data.frame}} or a 
\code{\link[base]{matrix}} containing ternary data-points.}

\item{\dots}{Additional parameters passed to \code{\link[graphics]{points}}.}
}
\value{
Invisibly returns the graphical element as x-y coordinates.
}
\description{
Add points to a ternary plot
}
\seealso{
\code{\link[graphics]{points}}.
}

