% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ternaryPolygons.R
\name{ternaryPolygons}
\alias{ternaryPolygons}
\alias{ternaryPolygons.ternaryPolygons}
\alias{ternaryPolygons.ternarySystem}
\title{Add polygons (a grid or ternary-classes) to an existing ternary plot}
\usage{
ternaryPolygons(s, ...)

\method{ternaryPolygons}{ternaryPolygons}(s, x = NULL, density = NULL,
  angle = 45, border = NULL, bg = NULL, lty = par("lty"), lwd = NULL,
  labels = NULL, adj = NULL, pos = NULL, offset = 0.5, vfont = NULL,
  cex = 1, col = NULL, font = NULL, ...)

\method{ternaryPolygons}{ternarySystem}(s, border = NULL, col = NULL,
  bg = NULL, lwd = NULL, ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternaryPolygons}}-object,
 such as created by \code{\link[ternaryplot]{createTernaryGrid}}
 or by \code{\link[ternaryplot]{ternaryClasses}}.}

\item{x}{Not used.}

\item{density}{See \code{\link[graphics]{polygon}}.}

\item{angle}{See \code{\link[graphics]{polygon}}.}

\item{border}{See \code{\link[graphics]{polygon}}.}

\item{bg}{Passed to \code{col} in \code{\link[graphics]{polygon}}.}

\item{lty}{See \code{\link[graphics]{polygon}}.}

\item{lwd}{See \code{\link[graphics]{par}}. Thickness of the polygon
 border.}

\item{labels}{Vector of character strings. Labels to be drawn on top
 of the polygons. If \code{NULL}, the default label
 in \code{s[[ 'labels' ]]} are used. If a single character
 string, recycled for each polygon. Set to \code{NA} to
 prevent any label to be drawn.}

\item{adj}{See \code{\link[graphics]{text}} (parameter of the text
 \code{labels}).}

\item{pos}{See \code{\link[graphics]{text}} (parameter of the text
 \code{labels}).}

\item{offset}{See \code{\link[graphics]{text}} (parameter of the text
 \code{labels}).}

\item{vfont}{See \code{\link[graphics]{text}} (parameter of the text
\code{labels}).}

\item{cex}{See \code{\link[graphics]{text}} (parameter of the text
 \code{labels}).}

\item{col}{See \code{\link[graphics]{text}} (parameter of the text
 \code{labels}).}

\item{font}{See \code{\link[graphics]{text}} (parameter of the text
 \code{labels}).}

\item{\dots}{Additional arguments passed to \code{\link[graphics]{polygon}}.


return
 Invisibly returns a list of polygon coordinates in
 xy coordinates.}
}
\description{
Add polygons (a grid or ternary-classes) to an existing ternary
 plot
}
\seealso{
\code{\link[ternaryplot]{ternaryPlot}} (or
 \code{\link[ternaryplot]{ternaryWindow}}) to create
 a plot (window) on which the ternary polygons can
 be added. \code{\link[graphics]{polygon}} for the equivalent
 non-ternary function.
}

