% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa02-ternaryplot-classes.R
\name{ternarySystem-class}
\alias{ternarySystem-class}
\title{Class to hold the definition of a ternarySystem (the full definition of ternary plot and optionally classification).}
\description{
Class (\code{S3}) to hold the definition of a ternarySystem 
 (the full definition of ternary plot and optionally 
 classification). A \code{ternarySystem-class} object is 
 a tagged \code{\link[base]{list}} of several 
 objects/items (see below). A 'tagged' list means a list of 
 the form \code{list( "tag1" = value1, "tag2" = value2, 
 ... )}.
}
\details{
The \code{S3} \code{\link[base]{class}} system is an 
 \emph{in}formal and lightweight class system.
}
\section{List objects/items}{

\itemize{
 \item \bold{\code{ternaryGeometry}}: See 
   \code{\link[ternaryplot]{ternaryGeometry-class}}.
 \item \bold{\code{ternaryVariables}}: See 
   \code{\link[ternaryplot]{ternaryVariables-class}}.
 \item \bold{\code{main}}: Single character string. 
   Title of the triangle plot.
 \item \bold{\code{vertices}}: See 
   \code{\link[ternaryplot]{tpPar}}. If non-null, 
   \code{\link[base]{data.frame}} with 4 columns: \code{id}, 
   and 3 other columns corresponding to \code{blrNames} 
   in \code{ternaryVariables}. If \code{NULL}, default 
   values will be used \code{getTpPar("vertices")}, and 
   the columns names changed to those of 
   \code{ternaryVariables}.
 \item \bold{\code{classes}}: See 
   \code{\link[ternaryplot]{tpPar}}. Please keep in mind 
   that the order of the class matters regarding the 
   classification of point ternary data, in the case of 
   "tights" (points that are in between two or more 
   classes): the function \code{ternaryClassify} uses 
   internally the function \code{\link[sp]{over}}, for 
   which the last polygon in which a point falls is 
   returned.
 \item \bold{\code{scale}}: NOT USED (YET). See 
   \code{\link[ternaryplot]{tpPar}}. If non-null, 
   \code{\link[base]{data.frame}} with 3 columns, 
   corresponding to \code{blrNames} in 
   \code{ternaryVariables}. If \code{NULL}, default values 
   will be used \code{getTpPar("scale")}, and the 
   columns names changed to those of \code{ternaryVariables}.
 \item \bold{\code{over}}: Either \code{NULL} or a 
   \code{\link{function}} with 3 arguments, \code{s}, 
   code{x} and \code{scale}. If a \code{\link{function}}, 
   should be used to add arbitrary graphical overlay on 
   top of ternary plots. Experimental.
}
}
\seealso{
\itemize{
 \item \code{\link[ternaryplot]{getTernarySystem}} to fetch 
   an existing (pre defined) ternary system, or 
   \code{\link[ternaryplot]{createTernarySystem}} to create 
   one. Notice that \code{\link[ternaryplot]{ternaryPlot}} 
   also output a \code{ternarySystem-class} object.
 \item \code{\link[ternaryplot]{ternaryPlot}} for plotting 
   a ternary system (with its ternary classification, when 
   relevant, and optionally with ternary data points), 
   \code{\link[ternaryplot]{ternaryClasses}} for extracting 
   the classes of a \code{ternarySystem-class} (as 
   \code{ternaryPolygons}) and \code{ternaryClassify} 
   for classifying ternary data.
 \item \code{\link[ternaryplot]{ternaryGeometry-class}} and 
   \code{\link[ternaryplot]{ternaryVariables-class}}, 
   two essentials building blocks of ternary systems.
 \item \code{\link[ternaryplot]{blrNames}} for fetching 
   or setting the variables names and 
   \code{\link[ternaryplot]{blrLabels}} for fetching 
   or setting the axis labels.
 \item \code{\link[ternaryplot]{tlrAngles}} for fetching 
   or setting the geometry's angle, 
   \code{\link[ternaryplot]{blrClock}} for fetching 
   or setting the geometry axis' directions and 
   \code{\link[ternaryplot]{fracSum}} for fetching 
   or setting the geometry's sum of fractions.
 \item \code{\link[ternaryplot]{ternaryCheck}}  
   for checking the validity of a 
   \code{ternarySystem-class} object.
}
}

