% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa03-ternaryplot-classes-utility.R
\name{ternarySystem}
\alias{ternarySystem}
\alias{ternarySystem.ternaryPolygons}
\title{Fetch the ternary system definition from another object}
\usage{
ternarySystem(x, ...)

\method{ternarySystem}{ternaryPolygons}(x, ...)
}
\arguments{
\item{x}{A \code{ternaryPolygons}-object, as
 created with \code{\link[ternaryplot]{createTernaryGrid}}.
 or \code{\link[ternaryplot]{ternaryClasses}}.}

\item{\dots}{Not used}
}
\value{
A \code{ternarySystem}-object.
}
\description{
Fetch the ternary system definition from another object
}

