% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa03-ternaryplot-classes-utility.R
\name{ternarySystem}
\alias{ternarySystem}
\alias{ternarySystem.ternaryPolygons}
\alias{ternarySystem<-}
\alias{ternarySystem<-.ternaryPolygons}
\title{Fetch or set the ternary system definition from another object}
\usage{
ternarySystem(s, ...)

\method{ternarySystem}{ternaryPolygons}(s, ...)

ternarySystem(s, ...) <- value

\method{ternarySystem}{ternaryPolygons}(s, ...) <- value
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternaryPolygons-class}} object.}

\item{value}{A \code{\link[ternaryplot]{ternarySystem-class}} object. 
The \code{ternarySystem-class} in \code{s} will be 
replaced by the one in \code{value}.}

\item{\dots}{Not used}
}
\value{
The \code{\link[ternaryplot]{ternarySystem-class}} 
 extracted from \code{s} or, 
 when using the \code{`<-`} method, a 
 \code{\link[ternaryplot]{ternaryPolygons-class}} object 
 with an updated \code{\link[ternaryplot]{ternarySystem-class}}.
}
\description{
Fetch or set the ternary system definition from another object
}

