% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa03-ternaryplot-classes-utility.R
\name{ternaryVariables}
\alias{ternaryVariables}
\alias{ternaryVariables.ternarySystem}
\alias{ternaryVariables<-}
\alias{ternaryVariables<-.ternarySystem}
\title{Set or get the ternaryVariables of a ternarySystem object.}
\usage{
ternaryVariables(s, ...)

\method{ternaryVariables}{ternarySystem}(s, ...)

ternaryVariables( s, ... ) <- value

\method{ternaryVariables}{ternarySystem}( s, ... ) <- value
}
\arguments{
\item{s}{A \code{ternarySystem} object, as created with
 \code{\link[ternaryplot]{createTernarySystem}}.}

\item{value}{A \code{ternaryVariables} object, as created with
 \code{\link[ternaryplot]{createTernaryVariables}}.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\description{
Set or get the ternaryVariables of a \code{ternarySystem} object.
}

