% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{ternaryWindow}
\alias{ternaryWindow}
\alias{ternaryWindow.character}
\alias{ternaryWindow.ternarySystem}
\title{Set up World Coordinates for a ternary plot (invisible base plot)}
\usage{
ternaryWindow(s, ...)

\method{ternaryWindow}{character}(s, x, scale = FALSE, ...)

\method{ternaryWindow}{ternarySystem}(s, x, scale = FALSE, ...)
}
\arguments{
\item{s}{A \code{ternarySystem} object, as created with
 \code{\link[ternaryplot]{createTernarySystem}}, or a single
 \code{character} string. Can be missing.}

\item{x}{A \code{\link[base]{data.frame}} or a \code{\link[base]{matrix}}
 containing point ternary data (x-y-x) to be ploted on the graph.
 It should contain the 3 columns names given in \code{s}. Can
 be missing or \code{NULL}.}

\item{scale}{A \code{\link[base]{data.frame}} with 3 columns and 2 rows.
 Contains the min and max limits of each of the 3 variables
 (columns = variables, rows = min and max).}

\item{\dots}{Additional parameters passed to specific methods.}
}
\description{
Set up World Coordinates for a ternary plot (invisible base plot)
}

