\docType{methods}
\name{tlrAngles}
\alias{tlrAngles}
\alias{tlrAngles,ternarySystem-method}
\alias{tlrAngles<-}
\alias{tlrAngles<-,ternarySystem-method}
\title{Set or get the top, left and right angles of a ternary system
 object}
\usage{
  \S4method{tlrAngles}{ternarySystem}( x, \dots )

  \S4method{tlrAngles}{ternarySystem}( x, \dots ) <- value
}
\arguments{
  \item{x}{A \code{\linkS4class{ternarySystem}} object.}

  \item{\dots}{Additional parameters passed to specific
  methods.}

  \item{value}{Vector of three numerical values, summing to
  180. Top, left and right angles of the ternary system
  object.}
}
\description{
  Set or get the top, left and right angles of a
  \code{\linkS4class{ternarySystem}} object
}

