% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy2ternary.R
\name{xy2ternary}
\alias{xy2ternary}
\alias{xy2ternary.character}
\alias{xy2ternary.ternarySystem}
\title{Converts x-y coordinates into ternary point-data.}
\usage{
xy2ternary(s, data, xyNames = c("x", "y"), ...)

\method{xy2ternary}{character}(s, data, xyNames = c("x", "y"), ...)

\method{xy2ternary}{ternarySystem}(s, data, xyNames = c("x", "y"), ...)
}
\arguments{
\item{s}{A \code{\link[ternaryplot]{ternarySystem}} object or a 
character string naming an pre-defined 
\code{\link[ternaryplot]{ternarySystem-class}}.}

\item{data}{A \code{\link[base]{data.frame}} or a 
\code{\link[base]{matrix}} with x-y coordinates.}

\item{xyNames}{A vector of 2 character strings. Names of the \code{x}- 
and \code{y}-variables in table \code{data}.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
Returns a \code{\link[base]{data.frame}} with ternary 
 point-data. As many rows as \code{nrow(data)} and 3 columns, 
 one per ternary-variable (same names as \code{blrNames(s)}).
}
\description{
Converts x-y coordinates into ternary point-data (bottom, 
 left, right axis), according to the specification of a 
 \code{\link[ternaryplot]{ternarySystem-class}}
}
\examples{

library( "ternaryplot" ) 

hypres <- getTernarySystem( "hypres" )

xyData <- data.frame(
    "x" = c( 0, 100,   0, 100, 50 ), 
    "y" = c( 0, 100, 100,   0, 50 ) 
)   

tData <- xy2ternary( s = hypres, data = xyData, 
    yxNames = c( "x", "y" ) )

tData

    # # Visualise the result:

    # ternaryStyle( "publication" )
    # ternaryPlot( hypres )
    # points( x = xyData[, "x" ], y = xyData[, "y" ], 
        # pch = 1, cex = 2 )
    # notNeg <- apply( X = tData, MARGIN = 1, 
        # FUN = function(x){ !any(x < 0) } )
    # ternaryPoints( s = hypres, x = tData[ notNeg, ], 
        # pch = 3 ) 
}

